refactor: reogranize molecules components
