+++ src/components/molecules/task-input/TaskInput.vue
@@ -1,47 +1,35 @@
 <template>
-  <div class="w-full flex flex-row justify-center items-center">
-    <input-view
-      id="task-input"
-      name="task-input"
-      placeholder="add new task"
-      type="text"
-      :value="value"
-      :on-input="handleInput"
-      :on-enter="handleClick"
-    ></input-view>
-    <button-svg type="button" icon-name="plus-square-thin" :on-click="handleClick"></button-svg>
+  <div class="w-full flex flex-col justify-center items-center">
+    <div class="w-full flex flex-row justify-center items-center">
+      <input-view
+        id="task-input"
+        name="task-input"
+        placeholder="add new task"
+        type="text"
+        :on-input="handleInput"
+        classes="border w-full p-1 rounded-sm border-sky-800"
+      />
+      <!-- <button-svg type="button" icon-name="plus-square-thin" :on-click="handleClick" /> -->
+    </div>
   </div>
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import InputView from '../../atoms/inputs/input-text/InputView.vue';
-import ButtonSvg from '../../atoms/buttons/button-svg/ButtonSvg.vue';
+import InputView from '../../atoms/input/InputView.vue';
+// import ButtonSvg from '../../atoms/button/ButtonSvg.vue';
 
 export default defineComponent({
   name: 'TaskInput',
   components: {
     InputView,
-    ButtonSvg,
-  },
-
-  data(): { value: string } {
-    return {
-      value: '',
-    };
+    // ButtonSvg,
   },
 
   methods: {
     handleInput(e: Event) {
       if (e instanceof InputEvent && e.target instanceof HTMLInputElement) {
-        this.value = e.target.value;
-      }
-    },
-
-    handleClick() {
-      if (this.value) {
-        this.$store.commit('addTask', this.value);
-        this.value = '';
+        this.$emit('update-value', e.target.value);
       }
     },
   },
